Move tui elements to own submodule
